requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.8
Django==3.0.5
gunicorn==20.0.4
idna==2.8
pycparser==2.19
pyOpenSSL==19.1.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
